requirements.txt
pyBCRAdata==0.4.4
reflex==0.7.7
python-dotenv==1.1.0